[#98460872] refactored Power.isCharging into Dispatcher.isBatteryCharging to ease testability and to fix a NPE we were having on elemez because of the unclear API we exposed
